Misc/tax-return/challenge/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
requests>=2.28
pdfplumber>=0.10
python-multipart
